% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{add_to_queue}
\alias{add_to_queue}
\title{Wave-based queue management - Add task to queue}
\usage{
add_to_queue(task_id, plan)
}
\value{
Modified plan object
}
\description{
Wave-based queue management - Add task to queue
}
\keyword{internal}
