% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_base_image}
\alias{build_base_image}
\title{Build base Docker image with common dependencies}
\usage{
build_base_image(region)
}
\description{
Build base Docker image with common dependencies
}
\keyword{internal}
