% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_environment_image}
\alias{build_environment_image}
\title{Build environment image}
\usage{
build_environment_image(tag, region, use_public = NULL)
}
\arguments{
\item{tag}{Image tag}

\item{region}{AWS region}

\item{use_public}{Logical, use public ECR base image (default NULL = from config)}
}
\description{
Build environment image
}
\keyword{internal}
