% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-estimate.R
\name{calculate_predictions}
\alias{calculate_predictions}
\title{Calculate Cloud Performance Predictions}
\usage{
calculate_predictions(
  n_total,
  avg_time_per_task,
  local_specs,
  workers,
  cpu,
  memory,
  platform
)
}
\description{
Calculate Cloud Performance Predictions
}
\keyword{internal}
