% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{check_and_submit_wave}
\alias{check_and_submit_wave}
\title{Check and Submit Wave if Ready}
\usage{
check_and_submit_wave(backend)
}
\arguments{
\item{backend}{Backend environment}
}
\description{
Checks wave queue and submits next wave if current wave is complete
}
\keyword{internal}
