% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_aws_credentials}
\alias{check_aws_credentials}
\title{Check AWS credentials}
\usage{
check_aws_credentials()
}
\value{
Logical indicating if credentials are valid
}
\description{
Check AWS credentials
}
\keyword{internal}
