% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ecr_image_age}
\alias{check_ecr_image_age}
\title{Check ECR image age and suggest/force rebuild}
\usage{
check_ecr_image_age(region, image_tag, ttl_days = NULL, force_rebuild = FALSE)
}
\arguments{
\item{region}{AWS region}

\item{image_tag}{Image tag to check}

\item{ttl_days}{TTL setting (NULL = no check)}

\item{force_rebuild}{Force rebuild if past TTL}
}
\value{
TRUE if image is fresh or doesn't exist, FALSE if stale
}
\description{
Check ECR image age and suggest/force rebuild
}
\keyword{internal}
