% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{check_fargate_quota}
\alias{check_fargate_quota}
\title{Check Fargate vCPU quota}
\usage{
check_fargate_quota(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
List with quota information
}
\description{
Check Fargate vCPU quota
}
\keyword{internal}
