% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-api.R
\name{cleanup_session}
\alias{cleanup_session}
\title{Cleanup session}
\usage{
cleanup_session(session, stop_workers = TRUE, force = FALSE)
}
\arguments{
\item{session}{Session object}

\item{stop_workers}{Stop running ECS tasks (default TRUE)}

\item{force}{Delete S3 files (default FALSE)}
}
\description{
Cleanup session
}
\keyword{internal}
