% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-api.R
\name{collect_session_results}
\alias{collect_session_results}
\title{Collect results from session}
\usage{
collect_session_results(session, wait, timeout)
}
\description{
Collect results from session
}
\keyword{internal}
