% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ecr_lifecycle_policy}
\alias{create_ecr_lifecycle_policy}
\title{Create ECR lifecycle policy to auto-delete old images}
\usage{
create_ecr_lifecycle_policy(region, repository_name, ttl_days = NULL)
}
\arguments{
\item{region}{AWS region}

\item{repository_name}{ECR repository name}

\item{ttl_days}{Number of days to keep images (NULL = no auto-delete)}
}
\description{
Create ECR lifecycle policy to auto-delete old images
}
\keyword{internal}
