% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-api.R
\name{create_session_object}
\alias{create_session_object}
\title{Create session object with methods}
\usage{
create_session_object(backend)
}
\arguments{
\item{backend}{Backend environment}
}
\value{
Session object (environment)
}
\description{
Create session object with methods
}
\keyword{internal}
