% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_starburst_bucket}
\alias{create_starburst_bucket}
\title{Create S3 bucket for staRburst}
\usage{
create_starburst_bucket(bucket_name, region)
}
\description{
Create S3 bucket for staRburst
}
\keyword{internal}
