% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_base_image}
\alias{ensure_base_image}
\title{Ensure base image exists}
\usage{
ensure_base_image(region, use_public = NULL)
}
\arguments{
\item{region}{AWS region}

\item{use_public}{Logical, use public ECR base image (default TRUE)}
}
\description{
Ensure base image exists
}
\keyword{internal}
