% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{ensure_ecs_instance_profile}
\alias{ensure_ecs_instance_profile}
\title{Ensure ECS instance IAM profile exists}
\usage{
ensure_ecs_instance_profile(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Instance profile ARN
}
\description{
Ensure ECS instance IAM profile exists
}
\keyword{internal}
