% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{ensure_ecs_security_group}
\alias{ensure_ecs_security_group}
\title{Ensure ECS security group exists}
\usage{
ensure_ecs_security_group(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Security group ID
}
\description{
Ensure ECS security group exists
}
\keyword{internal}
