% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{future_starburst}
\alias{future_starburst}
\title{Create a future using staRburst backend}
\usage{
future_starburst(
  expr,
  envir = parent.frame(),
  substitute = TRUE,
  globals = TRUE,
  packages = NULL,
  ...
)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{envir}{Environment for evaluation}

\item{substitute}{Substitute expression}

\item{globals}{List of global variables}

\item{packages}{Packages to load}

\item{...}{Additional arguments}
}
\description{
Create a future using staRburst backend
}
\keyword{internal}
