% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_architecture_from_instance_type}
\alias{get_architecture_from_instance_type}
\title{Get CPU architecture from instance type}
\usage{
get_architecture_from_instance_type(instance_type)
}
\arguments{
\item{instance_type}{EC2 instance type (e.g., "c7g.xlarge", "c7i.xlarge", "c7a.xlarge")}
}
\value{
CPU architecture ("ARM64" or "X86_64")
}
\description{
Get CPU architecture from instance type
}
\keyword{internal}
