% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{get_autoscaling_client}
\alias{get_autoscaling_client}
\title{Get Auto Scaling client}
\usage{
get_autoscaling_client(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Auto Scaling client
}
\description{
Get Auto Scaling client
}
\keyword{internal}
