% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_base_image_source}
\alias{get_base_image_source}
\title{Get base image source URI}
\usage{
get_base_image_source(use_public = TRUE)
}
\arguments{
\item{use_public}{Logical, use public ECR base image (default TRUE)}
}
\description{
Get base image source URI
}
\keyword{internal}
