% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{get_ec2_ondemand_price}
\alias{get_ec2_ondemand_price}
\title{Look up current On-Demand price via the AWS Pricing API.}
\usage{
get_ec2_ondemand_price(instance_type, region)
}
\description{
Look up current On-Demand price via the AWS Pricing API.
}
\keyword{internal}
