% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{get_ec2_spot_price}
\alias{get_ec2_spot_price}
\title{Look up recent Spot price via EC2 spot-price history (avg across AZs).}
\usage{
get_ec2_spot_price(instance_type, region)
}
\description{
Look up recent Spot price via EC2 spot-price history (avg across AZs).
}
\keyword{internal}
