% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{get_ecs_optimized_ami}
\alias{get_ecs_optimized_ami}
\title{Get ECS-optimized AMI ID for region and architecture}
\usage{
get_ecs_optimized_ami(region, architecture = "X86_64")
}
\arguments{
\item{region}{AWS region}

\item{architecture}{CPU architecture ("X86_64" or "ARM64")}
}
\value{
AMI ID
}
\description{
Get ECS-optimized AMI ID for region and architecture
}
\keyword{internal}
