% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_execution_role_arn}
\alias{get_execution_role_arn}
\title{Get IAM execution role ARN}
\usage{
get_execution_role_arn(region)
}
\description{
Returns the ARN for the ECS execution role (should be created during setup)
}
\keyword{internal}
