% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_instance_specs}
\alias{get_instance_specs}
\title{Get instance specifications (vCPUs, memory)}
\usage{
get_instance_specs(instance_type)
}
\description{
Get instance specifications (vCPUs, memory)
}
\keyword{internal}
