% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_instance_vcpus}
\alias{get_instance_vcpus}
\title{Get vCPU count for instance type}
\usage{
get_instance_vcpus(instance_type)
}
\arguments{
\item{instance_type}{EC2 instance type}
}
\value{
Number of vCPUs
}
\description{
Get vCPU count for instance type
}
\keyword{internal}
