% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_or_create_task_definition}
\alias{get_or_create_task_definition}
\title{Get or create task definition}
\usage{
get_or_create_task_definition(plan)
}
\description{
Get or create task definition
}
\keyword{internal}
