% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{get_pool_status}
\alias{get_pool_status}
\title{Get pool status}
\usage{
get_pool_status(backend)
}
\arguments{
\item{backend}{Backend configuration object}
}
\value{
List with pool status information
}
\description{
Query current state of the EC2 pool
}
\keyword{internal}
