% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_service_quotas_client}
\alias{get_service_quotas_client}
\title{Get Service Quotas client}
\usage{
get_service_quotas_client(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Service Quotas client
}
\description{
Get Service Quotas client
}
\keyword{internal}
