% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-state.R
\name{get_session_manifest}
\alias{get_session_manifest}
\title{Get session manifest from S3}
\usage{
get_session_manifest(session_id, region, bucket)
}
\arguments{
\item{session_id}{Session identifier}

\item{region}{AWS region}

\item{bucket}{S3 bucket}
}
\value{
Session manifest list
}
\description{
Get session manifest from S3
}
\keyword{internal}
