% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_starburst_config}
\alias{get_starburst_config}
\title{Get staRburst configuration}
\usage{
get_starburst_config()
}
\value{
List of configuration values
}
\description{
Get staRburst configuration
}
\keyword{internal}
