% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_starburst_security_groups}
\alias{get_starburst_security_groups}
\title{Get starburst security groups}
\usage{
get_starburst_security_groups(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Vector of security group IDs
}
\description{
Get starburst security groups
}
\keyword{internal}
