% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_starburst_subnets}
\alias{get_starburst_subnets}
\title{Get starburst subnets}
\usage{
get_starburst_subnets(region)
}
\arguments{
\item{region}{AWS region}
}
\value{
Vector of subnet IDs
}
\description{
Get starburst subnets
}
\keyword{internal}
