% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_task_role_arn}
\alias{get_task_role_arn}
\title{Get IAM task role ARN}
\usage{
get_task_role_arn(region)
}
\description{
Returns the ARN for the ECS task role (should be created during setup)
}
\keyword{internal}
