% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R, R/starburst-map.R,
%   R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b

a \%||\% b

a \%||\% b
}
\description{
Null-coalescing operator

Null-coalescing operator

Infix null-coalesce operator
}
\keyword{internal}
