% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-backend.R
\name{launch_detached_workers}
\alias{launch_detached_workers}
\title{Launch workers for detached session}
\usage{
launch_detached_workers(backend)
}
\arguments{
\item{backend}{Backend environment}
}
\value{
Invisibly returns NULL
}
\description{
Launches workers with bootstrap tasks that tell them the session ID
}
\keyword{internal}
