% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarburstBackend-class.R
\name{launchFuture.StarburstBackend}
\alias{launchFuture.StarburstBackend}
\title{Launch a future on the Starburst backend}
\usage{
\method{launchFuture}{StarburstBackend}(backend, future, ...)
}
\arguments{
\item{backend}{A StarburstBackend object}

\item{future}{The future object to launch}

\item{...}{Additional arguments}
}
\value{
The future object (invisibly)
}
\description{
Launch a future on the Starburst backend
}
