% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarburstBackend-class.R
\name{listFutures.StarburstBackend}
\alias{listFutures.StarburstBackend}
\title{List futures for StarburstBackend}
\usage{
\method{listFutures}{StarburstBackend}(backend, ...)
}
\arguments{
\item{backend}{A StarburstBackend object}

\item{...}{Additional arguments}
}
\value{
List of futures (empty for this backend)
}
\description{
List futures for StarburstBackend
}
