% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{package_installation_error}
\alias{package_installation_error}
\title{Package installation error}
\usage{
package_installation_error(package, error_message = NULL)
}
\description{
Package installation error
}
\keyword{internal}
