% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{parse_memory}
\alias{parse_memory}
\title{Parse memory specification}
\usage{
parse_memory(memory)
}
\arguments{
\item{memory}{Memory specification (numeric GB or string like "8GB")}
}
\description{
Parse memory specification
}
\keyword{internal}
