% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{permission_error}
\alias{permission_error}
\title{Permission denied error}
\usage{
permission_error(service, operation, resource = NULL, iam_role = NULL)
}
\description{
Permission denied error
}
\keyword{internal}
