% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-estimate.R
\name{print_estimate_summary}
\alias{print_estimate_summary}
\title{Print Estimate Summary}
\usage{
print_estimate_summary(pred, local_specs)
}
\description{
Print Estimate Summary
}
\keyword{internal}
