% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{quota_error}
\alias{quota_error}
\title{Quota exceeded error}
\usage{
quota_error(resource, requested, available, region, workers = NULL, cpu = NULL)
}
\description{
Quota exceeded error
}
\keyword{internal}
