% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-backend.R
\name{reconstruct_backend_from_manifest}
\alias{reconstruct_backend_from_manifest}
\title{Reconstruct backend from session manifest}
\usage{
reconstruct_backend_from_manifest(manifest)
}
\arguments{
\item{manifest}{Session manifest from S3}
}
\value{
Backend environment
}
\description{
Used when reattaching to an existing session
}
\keyword{internal}
