% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{request_quota_increase}
\alias{request_quota_increase}
\title{Request quota increase}
\usage{
request_quota_increase(service, quota_code, desired_value, region, reason = "")
}
\arguments{
\item{service}{AWS service (e.g., "fargate")}

\item{quota_code}{Service quota code}

\item{desired_value}{Desired quota value}

\item{region}{AWS region}

\item{reason}{Justification for increase}
}
\value{
Case ID if successful, NULL if failed
}
\description{
Request quota increase
}
\keyword{internal}
