% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{resolved.starburst_future}
\alias{resolved.starburst_future}
\title{Check if future is resolved}
\usage{
resolved.starburst_future(future, ...)
}
\arguments{
\item{future}{A starburst future object}

\item{...}{Additional arguments}
}
\value{
Logical indicating if resolved
}
\description{
Check if future is resolved
}
\keyword{internal}
