% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{resolved.StarburstFuture}
\alias{resolved.StarburstFuture}
\title{Check if StarburstFuture is Resolved}
\usage{
\method{resolved}{StarburstFuture}(future, ...)
}
\arguments{
\item{future}{A StarburstFuture object}

\item{...}{Additional arguments}
}
\value{
Logical indicating if the future is resolved
}
\description{
Checks whether the future task has completed execution
}
