% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{result.StarburstFuture}
\alias{result.StarburstFuture}
\title{Get Result from StarburstFuture}
\usage{
result.StarburstFuture(future, ...)
}
\arguments{
\item{future}{A StarburstFuture object}

\item{...}{Additional arguments}
}
\value{
A FutureResult object
}
\description{
Retrieves the result from a resolved future
}
