% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{run.StarburstFuture}
\alias{run.StarburstFuture}
\title{Run a StarburstFuture}
\usage{
run.StarburstFuture(future, ...)
}
\arguments{
\item{future}{A StarburstFuture object}

\item{...}{Additional arguments}
}
\value{
The future object (invisibly)
}
\description{
Submits the future task to AWS Fargate for execution
}
