% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{save_quota_request}
\alias{save_quota_request}
\title{Save quota request to local history}
\usage{
save_quota_request(case_id, service, quota_code, desired_value, region)
}
\description{
Save quota request to local history
}
\keyword{internal}
