% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-package.R
\docType{package}
\name{starburst-package}
\alias{starburst-package}
\title{starburst: Seamless AWS Cloud Bursting for Parallel R Workloads}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A future backend that enables seamless execution of parallel R workloads on AWS Fargate. staRburst handles environment synchronization, data transfer, quota management, and worker orchestration automatically, allowing users to scale from local execution to 100+ cloud workers with a single line of code change.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://starburst.ing}
  \item \url{https://github.com/scttfrdmn/starburst}
  \item Report bugs at \url{https://github.com/scttfrdmn/starburst/issues}
}

}
\author{
\strong{Maintainer}: Scott Friedman \email{help@starburst.ing}

}
\keyword{internal}
