% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarburstBackend-class.R
\name{starburst}
\alias{starburst}
\title{Starburst strategy marker}
\usage{
starburst(...)
}
\arguments{
\item{...}{Arguments passed to StarburstBackend()}
}
\value{
Does not return a value; always signals an error if called directly.
  This object exists as a strategy marker for \code{\link[future]{plan}}.
}
\description{
This function should never be called directly. Use plan(starburst, ...) instead.
}
