% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{starburst_check_quota_request}
\alias{starburst_check_quota_request}
\title{Monitor quota increase request}
\usage{
starburst_check_quota_request(case_id, region = NULL)
}
\arguments{
\item{case_id}{Case ID from quota increase request}

\item{region}{AWS region}
}
\value{
Invisibly returns the quota request details, or \code{NULL} on error.
}
\description{
Monitor quota increase request
}
\examples{
\donttest{
starburst_check_quota_request("case-12345")
}
}
