% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{starburst_cleanup_ecr}
\alias{starburst_cleanup_ecr}
\title{Clean up staRburst ECR images}
\usage{
starburst_cleanup_ecr(force = FALSE, region = NULL)
}
\arguments{
\item{force}{Delete all images immediately, ignoring TTL}

\item{region}{AWS region (default: from config)}
}
\description{
Manually delete Docker images from ECR to save storage costs.
Images will be rebuilt on next use (adds 3-5 min delay).
}
\examples{
\dontrun{
# Delete images past TTL
starburst_cleanup_ecr()

# Delete all images immediately (save $0.50/month)
starburst_cleanup_ecr(force = TRUE)
}
}
