% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-map.R
\name{starburst_cluster_map}
\alias{starburst_cluster_map}
\title{Execute Map on Starburst Cluster}
\usage{
starburst_cluster_map(cluster, .x, .f, .progress = TRUE)
}
\description{
Internal function to execute parallel map by creating StarburstFuture objects directly
}
\keyword{internal}
