% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{starburst_status}
\alias{starburst_status}
\title{Show staRburst status}
\usage{
starburst_status()
}
\value{
Invisibly returns a list with current configuration and quota information.
}
\description{
Show staRburst status
}
