% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2-pool.R
\name{stop_warm_pool}
\alias{stop_warm_pool}
\title{Stop warm pool}
\usage{
stop_warm_pool(backend)
}
\arguments{
\item{backend}{Backend configuration object}
}
\value{
Invisible NULL
}
\description{
Scales Auto-Scaling Group to zero
}
\keyword{internal}
