% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-backend.R
\name{submit_detached_worker}
\alias{submit_detached_worker}
\title{Submit detached worker to ECS}
\usage{
submit_detached_worker(backend, task_id)
}
\arguments{
\item{backend}{Backend environment}

\item{task_id}{Task ID for the worker to execute}
}
\value{
Task ARN
}
\description{
Submit detached worker to ECS
}
\keyword{internal}
