% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{submit_fargate_task}
\alias{submit_fargate_task}
\title{Submit Fargate Task}
\usage{
submit_fargate_task(future, backend)
}
\arguments{
\item{future}{StarburstFuture object}

\item{backend}{Backend/plan object}
}
\description{
Internal function to submit a task to ECS Fargate
}
\keyword{internal}
