% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-starburst.R
\name{submit_task}
\alias{submit_task}
\title{Submit Task to ECS}
\usage{
submit_task(future, backend)
}
\arguments{
\item{future}{StarburstFuture object}

\item{backend}{Backend/plan object}
}
\description{
Internal function to submit a task to ECS (Fargate or EC2)
}
\keyword{internal}
