% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-api.R
\name{submit_to_session}
\alias{submit_to_session}
\title{Submit task to session}
\usage{
submit_to_session(session, expr, envir, substitute, globals, packages)
}
\description{
Submit task to session
}
\keyword{internal}
