% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{suggest_quota}
\alias{suggest_quota}
\title{Suggest appropriate quota based on needs}
\usage{
suggest_quota(vcpus_needed)
}
\arguments{
\item{vcpus_needed}{vCPUs needed}
}
\value{
Suggested quota value
}
\description{
Suggest appropriate quota based on needs
}
\keyword{internal}
