% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-definition.R
\name{task-definition}
\alias{task-definition}
\title{ECS task definitions, IAM roles, and log groups for staRburst}
\description{
ECS task definitions, IAM roles, and log groups for staRburst
}
\keyword{internal}
