% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{task_failure_error}
\alias{task_failure_error}
\title{Task failure error}
\usage{
task_failure_error(task_id, reason = NULL, exit_code = NULL, log_stream = NULL)
}
\description{
Task failure error
}
\keyword{internal}
