% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-backend.R
\name{upload_detached_task}
\alias{upload_detached_task}
\title{Upload task for detached session}
\usage{
upload_detached_task(task_id, task_data, backend)
}
\arguments{
\item{task_id}{Task identifier}

\item{task_data}{Task data (expr, globals, packages, session_id)}

\item{backend}{Backend environment}
}
\value{
Invisibly returns NULL
}
\description{
Upload task for detached session
}
\keyword{internal}
