% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{value.starburst_future}
\alias{value.starburst_future}
\title{Get value from starburst future}
\usage{
value.starburst_future(future, ...)
}
\arguments{
\item{future}{A starburst future object}

\item{...}{Additional arguments}
}
\value{
The result of the future evaluation
}
\description{
Get value from starburst future
}
\keyword{internal}
