% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{worker_validation_error}
\alias{worker_validation_error}
\title{Worker validation error}
\usage{
worker_validation_error(workers, max_allowed = 500)
}
\description{
Worker validation error
}
\keyword{internal}
