% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-starburst.R
\name{check_and_submit_wave}
\alias{check_and_submit_wave}
\title{Check and submit wave if ready}
\usage{
check_and_submit_wave(plan)
}
\value{
Modified plan object
}
\description{
Check and submit wave if ready
}
\keyword{internal}
