% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-estimate.R, R/utils.R
\name{estimate_cost}
\alias{estimate_cost}
\title{Estimate Cost}
\usage{
estimate_cost(workers, cpu, memory, estimated_runtime_hours = 1)

estimate_cost(workers, cpu, memory, estimated_runtime_hours = 1)
}
\description{
Estimate Cost

Estimate cost
}
\keyword{internal}
