% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starburst-estimate.R, R/utils.R
\name{estimate_cost}
\alias{estimate_cost}
\title{Estimate Cost}
\usage{
estimate_cost(
  workers,
  cpu,
  memory,
  estimated_runtime_hours = 1,
  launch_type = "FARGATE",
  instance_type = NULL,
  use_spot = FALSE
)

estimate_cost(
  workers,
  cpu,
  memory,
  estimated_runtime_hours = 1,
  launch_type = "FARGATE",
  instance_type = NULL,
  use_spot = FALSE
)
}
\description{
Estimate Cost

Estimate cost
}
\keyword{internal}
