% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ec2_instance_price}
\alias{get_ec2_instance_price}
\title{Get EC2 instance pricing}
\usage{
get_ec2_instance_price(instance_type, use_spot = FALSE)
}
\arguments{
\item{instance_type}{EC2 instance type (e.g., "c7g.xlarge")}

\item{use_spot}{Whether to use spot pricing}
}
\value{
Price per hour in USD
}
\description{
Get EC2 instance pricing
}
\keyword{internal}
