% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarburstBackend-class.R
\name{nbrOfWorkers.StarburstBackend}
\alias{nbrOfWorkers.StarburstBackend}
\title{Number of workers for StarburstBackend}
\usage{
\method{nbrOfWorkers}{StarburstBackend}(evaluator)
}
\arguments{
\item{evaluator}{A StarburstBackend object}
}
\value{
Number of workers
}
\description{
Number of workers for StarburstBackend
}
