% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarburstBackend-class.R
\name{starburst}
\alias{starburst}
\title{Starburst strategy marker}
\usage{
starburst(...)
}
\arguments{
\item{...}{Arguments passed to StarburstBackend()}
}
\description{
This function should never be called directly. Use plan(starburst, ...) instead.
}
