% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{starburst_quota_status}
\alias{starburst_quota_status}
\title{Show quota status}
\usage{
starburst_quota_status(region = NULL)
}
\arguments{
\item{region}{AWS region (default: from config)}
}
\description{
Show quota status
}
\examples{
\dontrun{
starburst_quota_status()
}
}
