% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{starburst_rebuild_environment}
\alias{starburst_rebuild_environment}
\title{Rebuild environment image}
\usage{
starburst_rebuild_environment(region = NULL, force = FALSE)
}
\arguments{
\item{region}{AWS region (default: from config)}

\item{force}{Force rebuild even if current environment hasn't changed}
}
\description{
Rebuild environment image
}
\examples{
\dontrun{
starburst_rebuild_environment()
}
}
