% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{starburst_request_quota_increase}
\alias{starburst_request_quota_increase}
\title{Request quota increase (user-facing)}
\usage{
starburst_request_quota_increase(vcpus = 500, region = NULL)
}
\arguments{
\item{vcpus}{Desired vCPU quota}

\item{region}{AWS region (default: from config)}
}
\description{
Request quota increase (user-facing)
}
\examples{
\dontrun{
starburst_request_quota_increase(vcpus = 500)
}
}
