% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{starburst_setup}
\alias{starburst_setup}
\title{Setup staRburst}
\usage{
starburst_setup(region = "us-east-1", force = FALSE)
}
\arguments{
\item{region}{AWS region (default: "us-east-1")}

\item{force}{Force re-setup even if already configured}
}
\description{
One-time configuration to set up AWS resources for staRburst
}
\examples{
\dontrun{
starburst_setup()
}
}
